Restructure to allow for multiple command line verbs.
