+++ Options/RenderOptions.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Reflection;
 using CommandLine;
@@ -10,20 +11,21 @@ using RayTracer.Pixels;
 [assembly: AssemblyCopyright("Copyright \u00a9 2024")]
 [assembly: AssemblyInformationalVersion("1.0.1")]
 
-namespace RayTracer;
+namespace RayTracer.Options;
 
 /// <summary>
 /// This class represents the command line options that the user may specify to the ray
 /// tracer.
 /// </summary>
 // ReSharper disable once ClassNeverInstantiated.Global
-public class ProgramOptions
+public class RenderOptions
 {
     [Option('i', "input-file", Required = true,
         HelpText = "The name of the input file to process.")]
     public string InputFileName
     {
         get => _inputFileName;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             string path = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), value));
@@ -37,6 +39,7 @@ public class ProgramOptions
 
     [Option('d', "output-dir", Required = false,
         HelpText = "The name of the directory where the output file will be written.")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public string OutputDirectory
     {
         get => _outputDirectory;
@@ -56,6 +59,7 @@ public class ProgramOptions
     public string OutputFileName
     {
         get => GetOutputFileName();
+        // ReSharper disable once UnusedMember.Global
         set
         {
             string path = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), value));
@@ -66,6 +70,7 @@ public class ProgramOptions
 
     [Option('e', "output-extension", Required = false, SetName = "outputExtension",
         HelpText = "The name of the output file to write the rendered image to.")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public string OutputFileExtension
     {
         get => _outputFileExtension;
@@ -77,6 +82,7 @@ public class ProgramOptions
     public string OutputImageFormat
     {
         get => _outputImageFormat;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (!AvailableCodecs.IsFormatSupported(value))
@@ -91,6 +97,7 @@ public class ProgramOptions
     public int? Width
     {
         get => _width;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is < 1 or > 16384)
@@ -105,6 +112,7 @@ public class ProgramOptions
     public int? Height
     {
         get => _height;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is < 1 or > 16384)
@@ -119,6 +127,7 @@ public class ProgramOptions
     public int FrameRate
     {
         get => _frameRate;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is < 1)
@@ -133,6 +142,7 @@ public class ProgramOptions
     public long? Frame
     {
         get => _frame;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is < 0)
@@ -147,6 +157,7 @@ public class ProgramOptions
     public int BitsPerChannel
     {
         get => _bitsPerChannel;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is not 8 and not 16)
@@ -161,6 +172,7 @@ public class ProgramOptions
     public double? Gamma
     {
         get => _gamma;
+        // ReSharper disable once UnusedMember.Global
         set
         {
             if (value is < 0 or > 5)
@@ -175,24 +187,29 @@ public class ProgramOptions
     public bool? NoGamma
     {
         get => !_applyGamma;
+        // ReSharper disable once UnusedMember.Global
         set => _applyGamma = !value;
     }
 
     [Option("report-gamma", Required = false,
         HelpText = "If specified, the gamma correction value will be included in the image output file, if supported.")]
+    [SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global")]
     public bool? ReportGamma { get; set; }
 
     [Option("no-shadows", Required = false,
         HelpText = "Disable shadow rendering on all objects.")]
+    [SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global")]
     public bool? NoShadows { get; set; }
 
     [Option("grayscale", Required = false,
         HelpText = "Grayscale the image when written to image file.")]
+    [SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global")]
     public bool Grayscale { get; set; }
 
     [Option('l', "output-level", Required = false, Default = "normal",
         // ReSharper disable once StringLiteralTypo
         HelpText = "Sets the desired level of output.  Must be one of, [q]uiet, [n]ormal, [c]hatty or [v]erbose.  The values are not case-sensitive.")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public string OutputLevelText
     {
         get => OutputLevel.ToString().ToLowerInvariant();
@@ -202,10 +219,11 @@ public class ProgramOptions
     /// <summary>
     /// This property holds the output level the renderer is to use.
     /// </summary>
-    public OutputLevel OutputLevel { get; private set; }
+    public OutputLevel OutputLevel { get; private set; } = OutputLevel.Normal;
 
     [Option('a', "antialias", Required = false,
         HelpText = "Sets what sort of antialiasing should be applied to the image being rendered.")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public string AntiAliasingText
     {
         get => AntiAliasing.ToString();
@@ -215,36 +233,21 @@ public class ProgramOptions
     /// <summary>
     /// This property holds the antialiasing option for the ray tracer.
     /// </summary>
-    public AliasingOption AntiAliasing { get; }
+    public AliasingOption AntiAliasing { get; } = new();
 
     private string _inputFileName;
     private string _outputDirectory;
     private string _outputFileName;
     private string _outputFileExtension;
-    private string _outputImageFormat;
+    private string _outputImageFormat = "png";
     private int? _width;
     private int? _height;
-    private int _frameRate;
+    private int _frameRate = 24;
     private long? _frame;
-    private int _bitsPerChannel;
+    private int _bitsPerChannel = 8;
     private double? _gamma;
     private bool? _applyGamma;
 
-    public ProgramOptions()
-    {
-        _width = null;
-        _height = null;
-        _frameRate = 24;
-        _frame = null;
-        _bitsPerChannel = 8;
-        _gamma = null;
-        _applyGamma = null;
-        _outputImageFormat = "png";
-
-        OutputLevel = OutputLevel.Normal;
-        AntiAliasing = new AliasingOption();
-    }
-    
     /// <summary>
     /// This is a helper method for properly deriving the right output file name based on
     /// all the various things the user could have specified.
@@ -268,7 +271,8 @@ public class ProgramOptions
 
     /// <summary>
     /// This is a helper method for converting a piece of text to the output level it
-    /// represents.  We do so by making the input case-insensitive and allowed to be an
+    /// represents.
+    /// We do so by treating the input in a case-insensitive way and allow it to be an
     /// abbreviation.
     /// </summary>
     /// <param name="levelText">The text to start with.</param>
